+++ media/java/android/media/FaceDetector.java
@@ -102,6 +102,7 @@ public class FaceDetector {
      * Creates a FaceDetector, configured with the size of the images to
      * be analysed and the maximum number of faces that can be detected.
      * These parameters cannot be changed once the object is constructed.
+     * Note that the width of the image must be even.
      * 
      * @param width  the width of the image
      * @param height the height of the image
@@ -190,9 +191,9 @@ public class FaceDetector {
     native private void fft_get_face(Face face, int i);
     native private void fft_destroy();
 
-    private int     mFD;
-    private int     mSDK;
-    private int     mDCR;
+    private long    mFD;
+    private long    mSDK;
+    private long    mDCR;
     private int     mWidth;
     private int     mHeight;
     private int     mMaxFaces;
